refactor(ui): rename account screen
