+++ packages/extension/src/ui/App.tsx
@@ -9,7 +9,7 @@ import { useEntry } from "./hooks/useEntry"
 import { useTransactionErrorScreen } from "./hooks/useTransactionErrorScreen"
 import { routes } from "./routes"
 import { AccountListScreen } from "./screens/AccountListScreen"
-import { AccountScreen } from "./screens/AccountTokensScreen"
+import { AccountScreen } from "./screens/AccountScreen"
 import { ActionScreen } from "./screens/ActionScreen"
 import { AddTokenScreen } from "./screens/AddTokenScreen"
 import { BackupDownloadScreen } from "./screens/BackupDownloadScreen"
